requirements.txt
Django==5.2.5
django-cors-headers==4.4.0
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
yfinance==0.2.65